Seperate schedule-queue into plugin-owned
