+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+1.0.8 / 2019-12-27
+=================
+
+* Seperate schedule-task into plugin-owned
+
 1.0.8 / 2019-12-26
 =================
