+++ CTFd/plugins/ctfd_glowworm/control_utils.py
@@ -22,7 +22,10 @@ class ControlUtil:
     def remove_competition():
         mode = utils.get_config("user_mode")
         from .schedule import scheduler
-        scheduler.remove_job('time_base')
+        try:
+            scheduler.remove_job('time_base')
+        except Exception as e:
+            pass
         challenges = ADAChallenge.query.all()
         # 清除运行的容器
         for challenge in challenges:
